+++ src/schemas/database/project/slabType_v0.ts
@@ -1,10 +1,11 @@
 import { z } from 'zod'
+import { disableUserWrite } from '../../utility'
 
 export const SlabTypeDataSchema_v0 = z.object({
     color: z.string().optional(),
     fFactor: z.number().optional(),
     name: z.string().optional(),
-    timestamp: z.number().optional(),
+    timestamp: disableUserWrite(z.number().optional()),
     uValue: z.number().optional(),
     unconditionedCoolingTempF: z.number().optional(),
     unconditionedHeatingTempF: z.number().optional(),