+++ src/schemas/database/project/space_v0.ts
@@ -1,5 +1,6 @@
 import { z } from 'zod'
 import { LoadConditionSchema_v0 } from '../../misc'
+import { disableUserWrite } from '../../utility'
 import { ModeIdSchema_v0 } from './equipment_v0'
 
 const InfiltrationRequirementMethods = {
@@ -145,7 +146,7 @@ export const SpaceDataSchema_v0 = z.object({
     name: z.string().optional(),
     number: z.string().optional(),
     occupancy: z.number().optional(),
-    processed: z.boolean().optional(),
+    processed: disableUserWrite(z.boolean().optional()),
     revitId: z.string().optional(),
     roofAzimuth: z.number().optional(),
     roofPitch: z.number().optional(),
@@ -153,11 +154,11 @@ export const SpaceDataSchema_v0 = z.object({
     skylights: z.record(z.string(), SkylightDataSchema_v0).optional(),
     slabHeight: z.number().optional(),
     slabTypeId: z.string().optional(),
-    spaceNameAndNumberInputHash: z.string().optional(),
+    spaceNameAndNumberInputHash: disableUserWrite(z.string().optional()),
     spaceTypeId: z.string().optional(),
-    suggested: z.boolean().optional(),
-    suggestedSpaceName: z.string().optional(),
-    suggestedSpaceNumber: z.string().optional(),
+    suggested: disableUserWrite(z.boolean().optional()),
+    suggestedSpaceName: disableUserWrite(z.string().optional()),
+    suggestedSpaceNumber: disableUserWrite(z.string().optional()),
     zoneId: z.string().optional(),
 })
 export type SpaceData_v0 = z.infer<typeof SpaceDataSchema_v0>