+++ src/schemas/database/project/wallType_v0.ts
@@ -1,4 +1,5 @@
 import { z } from 'zod'
+import { disableUserWrite } from '../../utility'
 
 export const WallTypeDataSchema_v0 = z.object({
     ashraeWallTypeId: z.string().optional(),
@@ -7,7 +8,7 @@ export const WallTypeDataSchema_v0 = z.object({
     color: z.string().optional(),
     name: z.string().optional(),
     surfaceAbsorptance: z.number().optional(),
-    timestamp: z.number().optional(),
+    timestamp: disableUserWrite(z.number().optional()),
     uValue: z.number().optional(),
 })
 export type WallTypeData_v0 = z.infer<typeof WallTypeDataSchema_v0>